Mouse Selection
The grid keeps track of its origin and the screens origin, and can convert between screen points to cell points. Added mouse event where on mouse down I get the mouse position and select the appropriate grid square. If a cell is selected the grid draws it as a red rect.