Event System
Created a singleton class called EventSystem that I can subscribe events to anywhere in the program, and when I poll events in App it will call all subscribed methods for the specific event type. Also, made the seed for the noise be a random number between zero and one million.