Fall off Function
Created a fall off function, so as you go away from the center the height becomes much lower making an island. Added sand that is between water and grass. Fixed issue where Map texture was too small, and created a view window for which only a small portion of the map texture is shown on the screen, and w, a, s, and d keys can move the view window around on the map.