Refactored Map and MapGen
Recfactored MapGen to create the Map class which is just the Graph at this point. I create a MapGen in the App and call GenerateMap which returns a Map that has a reference to the MapGen if it needs access later for map width and height, and the seed. This reorganization is for the new Group class where I will group the terrain by common terrain type, and then use this for further generation like rivers.