BlitText
Created a BlitText that can take a font surface from the glyph atlas and blit it to another surface, for CPU side rendering. Then MapNode can draw to a surface, and in MapChunk I generate a surface draw all the MapNodes to it, and then create a texture from it. Fixed issue where ChunkRows and Cols was not being used in the MapChunk class.