UI Update
Removed UI from App. Added UI to both Player and Map Viewer, so that in the map view the group that the mouse is on shows up on the screen, and in the player view the group of the cell the player is currently in is shown on the screen.